requirements.txt
Django
requests
beautifulsoup4
html5lib
APScheduler
psycopg[binary]
crawlerdetect
django-tailwind
markdown
gunicorn
tweepy
g4f
pillow
google-api-python-client
whitenoise
sentry-sdk[django]
django-debug-toolbar
praw